Rename view to view-controller
